Remove some redundant use of `\textbf{}`
Using `\textbf{}` within another `\textbf{}` has no effect except
to make the source code harder to read.  Removing the duplicates
has no effect on the printed document.

Additionally, remove a leading space in the name of the first two
conferences that was causing them to be out of line with the rest.